2019 9

+++ src/solutions/year2019/day02.rs
@@ -1,17 +1,17 @@
 use crate::intcode::VM;
 
-fn run(mut vm: VM, noun: i32, verb: i32) -> i32 {
+fn run(mut vm: VM, noun: i64, verb: i64) -> i64 {
     vm.mem[1] = noun;
     vm.mem[2] = verb;
     vm.halt();
     vm.mem[0]
 }
 
-pub fn part1(input: &str) -> i32 {
+pub fn part1(input: &str) -> i64 {
     run(VM::new(input), 12, 2)
 }
 
-pub fn part2(input: &str) -> i32 {
+pub fn part2(input: &str) -> i64 {
     let vm = VM::new(input);
     for noun in 0..100 {
         for verb in 0..100 {
@@ -24,7 +24,7 @@ pub fn part2(input: &str) -> i32 {
 }
 
 pub fn tests() {
-    fn assert_mem(input: &str, expected: &[i32]) {
+    fn assert_mem(input: &str, expected: &[i64]) {
         let mut vm = VM::new(input);
         vm.halt();
         assert_eq!(&vm.mem[..expected.len()], expected);
